git-svn-id: http://caml.inria.fr/svn/ocaml/branches/contracts@11227 f963ae5c-01c2-4b8c-9fe0-0dff7051ff02

+++ tools/checkstack.c
@@ -35,7 +35,8 @@ int main(int argc, char ** argv)
              "Under sh, bash, zsh:  ulimit -s %lu\n"
              "Under csh, tcsh:      limit stacksize %lu\n\n",
              (unsigned long) (limit.rlim_cur / 1024),
-             MINSTACKBYTES / 1024, MINSTACKBYTES / 1024);
+             (unsigned long) (MINSTACKBYTES / 1024),
+             (unsigned long) (MINSTACKBYTES / 1024));
     exit (3);
   }
   exit (0);
